CIV-9249 refactor code and modified datatype of ErrorCode
