+++ src/main/java/uk/gov/hmcts/reform/civil/requestbody/CreateClaimCCD.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.mappings;
+package uk.gov.hmcts.reform.civil.requestbody;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
