+++ src/main/java/uk/gov/hmcts/reform/civil/service/SubmitCreateClaim.java
@@ -17,7 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.civil.mappings.CreateClaimCCD;
+import uk.gov.hmcts.reform.civil.requestbody.CreateClaimCCD;
 import uk.gov.hmcts.reform.civil.responsebody.CreateClaimErrorResponse;
 
 @Slf4j
